==> code/08-Cluster Location Info Scraping.R <==
# Scraping information to build predictive models. Note you'll need a Census API key to query some of this info. 
library(dplyr);library(tigris, options(tigris_use_cache = TRUE)); library(osmdata)
library(censusapi); library(readxl); library(tibble); library(terra)

#using only those counties we have data for
#calculating Hawaiian data, but not including in combined datasets as streamflow is missing

# City Vectors -----
cities.all = vect("maps/cb_2018_us_ua10_500k.shp")
## Missing Morristown NJ
## Need combining: San Fran San Jose and Cleveland Akron
cn = c("Albuquerque, NM", "Ann Arbor, MI", "Athens-Clarke County, GA",
       "Atlanta, GA", "Bellingham, WA", "Cedar City, UT",
       "Colorado Springs, CO", "Dallas--Fort Worth--Arlington, TX",
       "Denver--Aurora, CO", "Flagstaff, AZ", "Gainesville, FL", 
       "La Crosse, WI--MN", "Lawrence, KS", 
       "Los Angeles--Long Beach--Anaheim, CA", "Riverside--San Bernardino, CA",
       "Minneapolis--St. Paul, MN--WI", 
       "Nashville-Davidson, TN",
       "Phoenix--Mesa, AZ", "Portland, OR--WA", 
       "Salt Lake City--West Valley City, UT",
       "San Diego, CA", "San Francisco--Oakland, CA", "San Jose, CA",
       "San Marcos, TX",
       "Tampa--St. Petersburg, FL", "State College, PA", "Wooster, OH",
       "Cleveland, OH", "Akron, OH", "Youngstown, OH--PA")

cn = sort(cn)
cities = cities.all[cities.all$NAME10 %in% cn]

# Combine
cities$NAME10 = gsub("San Jose, CA", "San Francisco--Oakland, CA", cities$NAME10)
cities$NAME10 = gsub("Akron, OH", "Cleveland, OH", cities$NAME10)
cities$NAME10 = gsub("Riverside--San Bernardino, CA", "Los Angeles--Long Beach--Anaheim, CA", 
                     cities$NAME10)
cities = aggregate(cities, by = "NAME10")

#Drop unneeded fields
values(cities) = subset(values(cities), select = "NAME10")

# Add Morristown
MST = getbb("Morristown, NJ, USA", format_out = "polygon")
MST = vect(MST, crs = "WGS84", type = "polygon")
MST$NAME10 = "Morristown"
cities = rbind(cities, MST)

## Add our names
cities$NAME10 = gsub(",.*", "", cities$NAME10)
cities$NAME10 = gsub("Athens-Clarke County", "Athens", cities$NAME10)
cities$NAME10 = gsub("Cleveland", "Cleveland-Akron", cities$NAME10)
cities$NAME10 = gsub("Dallas.*", "Dallas Fort Worth", cities$NAME10)
cities$NAME10 = gsub("Denver.*", "Denver", cities$NAME10)
cities$NAME10 = gsub("Los Angeles.*", "Los Angeles", cities$NAME10)
cities$NAME10 = gsub("Minneapolis.*", "Minneapolis", cities$NAME10)
cities$NAME10 = gsub("Nashville.*", "Nashville", cities$NAME10)
cities$NAME10 = gsub("Phoenix.*", "Phoenix", cities$NAME10)
cities$NAME10 = gsub("Salt.*", "Salt Lake City", cities$NAME10)
cities$NAME10 = gsub("San Francisco.*", "San Francisco", cities$NAME10)
cities$NAME10 = gsub("Tampa.*", "St Petersburg", cities$NAME10)

#cities$total_area = round(expanse(cities, unit = "km"), 1)

tapData = read.csv("data/cityWater.csv")
tapData = vect(tapData, geom = c("lon", "lat"), crs = "WGS84")

for(i in seq_along(cities)){
  td = tapData[tapData$cluster_location == cities$NAME10[i]]
  ch = convHull(td)
  cities$total_area[i] = expanse(ch, unit = "km")
}

# Elevation, precip, streamflow -------------------------------

ruggedness <- rast("maps/elev_diff.tif")
cities <- project(cities, ruggedness) %>% crop(ruggedness)
df = terra::extract(ruggedness, cities, fun = max, na.rm = TRUE, ID = FALSE)
cities$ruggedness <- round(df, 0)

precip <- project(rast("maps/precip_mean.tif"), ruggedness)
streamflow <- project(rast("maps/streamflow_mean.tif"), ruggedness)

df <- terra::extract(streamflow, cities, fun = mean, na.rm = TRUE, ID = FALSE)
cities$streamflow <- df

df <- terra::extract(precip, cities, fun = mean, na.rm = TRUE, ID = FALSE)
cities$precip <- df

multivariate <- as.data.frame(cities) 
multivariate <- multivariate %>% 
  rename(cluster_location = NAME10)

# Census demographic info -------------------------------------------------

Sys.setenv(CENSUS_KEY = "YOUR KEY HERE")
# Reload .Renviron
readRenviron("~/.Renviron")
# Check to see that the expected key is output in your R console
Sys.getenv("CENSUS_KEY")

acs_simple <- getCensus(
  name = "acs/acs5",
  vintage = 2020,
  vars = c("NAME", "B01001_001E", "B19013_001E"),
  region = "county:*") %>% 
  rename(pop = B01001_001E, 
         medincome = B19013_001E)
acs_simple$GEOID = paste0(acs_simple$state, acs_simple$county)

# Creating Developed Areas Vectors ----------------------------------------

AA <- counties("MI", cb = T, resolution = "20m") %>% 
  filter(NAME %in% c("Washtenaw")) %>% 
  mutate(id = "Ann Arbor") %>% 
  vect() 

ABQ <- counties("NM", cb = T, resolution = "20m") %>% 
  filter(NAME %in% c("Bernalillo")) %>% 
  mutate(id = "Albuquerque") %>% 
  vect()

ATH <- counties("GA", cb = T, resolution = "20m") %>% 
  filter(NAME %in% c("Clarke","Oconee")) %>% 
  vect() %>% 
  terra::aggregate()
ATH$id <- "Athens"   

ATL <- counties("GA", cb = T, resolution = "20m") %>% 
  filter(NAME %in% c("Cobb", "DeKalb", "Fulton"))%>% 
  vect() %>% 
  terra::aggregate()
ATL$id <- "Atlanta"

BEL <- counties("WA", cb = T, resolution = "20m") %>% 
  filter(NAME %in% c("Whatcom"))%>% 
  mutate(id = "Bellingham") %>% 
  vect()

CED <- counties("UT", cb = T, resolution = "20m") %>% 
  filter(NAME %in% c("Iron")) %>%
  mutate(id = "Cedar City") %>% 
  vect()

CLE <- counties("OH", cb = T, resolution = "20m") %>% 
  filter(NAME %in% c("Cuyahoga", "Geauga", "Lake", "Medina")) %>% 
  vect() %>% 
  terra::aggregate()
CLE$id <- "Cleveland-Akron"

COL <- counties("CO", cb = T, resolution = "20m") %>% 
  filter(NAME %in% c("El Paso")) %>% 
  mutate(id = "Colorado Springs") %>% 
  vect()

DFW <- counties("TX", cb = T, resolution = "20m") %>% 
  filter(NAME %in% c("Collin County", "Dallas", "Ellis", "Johnson",
                     "Tarrant"))%>% 
  vect()%>% 
  terra::aggregate()
DFW$id <- "Dallas Fort Worth"

DEN <- counties("CO", cb = T, resolution = "20m") %>% 
  filter(NAME %in% c("Adams", "Arapahoe", "Boulder","Broomfield",
                     "Denver", "Jefferson")) %>% 
  vect() %>% 
  terra::aggregate()
DEN$id <- "Denver"

FLG <- counties("AZ", cb = T, resolution = "20m") %>% 
  filter(NAME %in% c("Coconino"))%>% 
  mutate(id = "Flagstaff") %>% 
  vect()

GNV <- counties("FL", cb = T, resolution = "20m") %>% 
  filter(NAME %in% c("Alachua"))%>%
  mutate(id = "Gainesville") %>% 
  vect()

LCR1 <- counties("MN", cb = T, resolution = "20m") %>% 
  filter(NAME %in% c("Houston", "Winona")) %>% 
  vect() 
LCR2 <- counties("WI", cb = T, resolution = "20m") %>% 
  filter(NAME %in% c("La Crosse"))  %>% 
  vect()
LCR <- terra::aggregate(terra::union(LCR1, LCR2))
rm(LCR1, LCR2)
LCR$id <- "La Crosse"

LAW <- counties("KS", cb = T, resolution = "20m") %>% 
  filter(NAME %in% c("Douglas"))%>% 
  mutate(id = "Lawrence") %>% 
  vect()

LAX <- counties("CA", cb = T, resolution = "20m") %>% 
  filter(NAME %in% c("Los Angeles", "Riverside",
                     "San Bernardino","San Diego"))%>%  
  vect() %>% 
  terra::aggregate()
LAX$id <- "Los Angeles"

MSP <- counties("MN", cb = T, resolution = "20m") %>% 
  filter(NAME %in% c("Anoka", "Hennepin", "Ramsey"))%>% 
  vect() %>% 
  terra::aggregate()
MSP$id <- "Minneapolis"

MOR <- counties("NJ", cb = T, resolution = "20m") %>% 
  filter(NAME %in% c("Essex", "Morris", "Somerset", "Union"))%>% 
  vect() %>% 
  terra::aggregate()
MOR$id <- "Morristown"

NAS <- counties("TN", cb = T, resolution = "20m") %>% 
  filter(NAME %in% c("Cheatham", "Davidson", "Rutherford", "Williamson"))%>% 
  vect() %>% 
  terra::aggregate()
NAS$id <- "Nashville"

PHX <- counties("AZ", cb = TRUE, resolution = "20m") %>% 
  filter(NAME %in% c("Maricopa", "Pinal")) %>% 
  vect() %>% 
  terra::aggregate()
PHX$id <- "Phoenix"

PTD1 <- counties("OR", cb = T, resolution = "20m") %>% 
  filter(NAME %in% c("Clackamas", "Multnomah")) %>% 
  vect()
PTD2 <- counties("WA", cb = T, resolution = "20m") %>% 
  filter(NAME %in% c("Clark")) %>% 
  vect()
PTD <- terra::aggregate(terra::union(PTD1, PTD2)) 
rm(PTD2, PTD1)
PTD$id <- "Portland"

SLC <- counties("UT", cb = TRUE, resolution = "20m") %>% 
  filter(NAME %in% c("Salt Lake", "Davis")) %>% 
  vect() %>% 
  terra::aggregate()
SLC$id <- 'Salt Lake City'

SD <- counties("CA", cb = T, resolution = "20m") %>% 
  filter(NAME %in% c("San Diego")) %>% 
  mutate(id = "San Diego") %>% 
  vect()

SF <- counties("CA", cb = T, resolution = "20m") %>% 
  filter(NAME %in% c("Alameda","Contra Costa", "Marin", "San Francisco",
                     "San Mateo","Santa Clara")) %>% 
  vect() %>% 
  terra::aggregate()
SF$id <- "San Francisco"

SM <- counties("TX", cb = T, resolution = "20m") %>% 
  filter(NAME %in% c("Hays"))%>% 
  mutate(id = "San Marcos") %>% 
  vect()

SP <- counties("FL", cb = T, resolution = "20m") %>% 
  filter(NAME %in% c("Pinellas"))%>% 
  mutate(id = "St Petersburg") %>% 
  vect()

SC <- counties("PA", cb = T, resolution = "20m") %>% 
  filter(NAME %in% c("Centre"))%>% 
  mutate(id = "State College") %>% 
  vect()

WOO <- counties("OH", cb = T, resolution = "20m") %>% 
  filter(NAME %in% c("Wayne"))%>% 
  mutate(id = "Wooster") %>% 
  vect()

YNG <- counties("OH", cb = T, resolution = "20m") %>% 
  filter(NAME %in% c("Mahoning", "Trumbull")) %>% 
  vect() %>% 
  terra::aggregate()
YNG$id = "Youngstown"

#Let's combine population and average median income. 
AA$pop <- subset(acs_simple, GEOID == "26161")$pop
AA$medincome <- subset(acs_simple, GEOID == "26161")$medincome

ABQ$pop <- subset(acs_simple, GEOID == "35001")$pop
ABQ$medincome <- subset(acs_simple, GEOID == "35001")$medincome

ATH$pop <- sum(subset(acs_simple, 
                      GEOID == "13219" | GEOID == "13059")$pop)
ATH$medincome <- weighted.mean(subset(acs_simple, 
                                      GEOID == "13219" | GEOID == "13059")$medincome,
                               subset(acs_simple, 
                                      GEOID == "13219" | GEOID == "13059")$pop)

ATL$pop <- sum(subset(acs_simple, 
                      GEOID == "13121" |
                        GEOID == "13089" | 
                        GEOID == "13067")$pop)
ATL$medincome <- weighted.mean(subset(acs_simple, 
                                      GEOID == "13121" |
                                        GEOID == "13089" | 
                                        GEOID == "13067")$medincome,
                               subset(acs_simple, 
                                      GEOID == "13121" |
                                        GEOID == "13089" | 
                                        GEOID == "13067")$pop)

BEL$pop <- subset(acs_simple, GEOID == "53073")$pop 
BEL$medincome <- subset(acs_simple, GEOID == "53073")$medincome 

CED$pop <- subset(acs_simple, GEOID == "49021")$pop 
CED$medincome <- subset(acs_simple, GEOID == "49021")$medincome 

# CLE counties "Cuyahoga", "Geauga", "Lake", "Medina"
CLE$pop <- sum(subset(acs_simple,
                      GEOID == "39035" | GEOID == "39055" | GEOID == "39085" | 
                        GEOID == "39103")$pop)
CLE$medincome <- weighted.mean(subset(acs_simple,
                                      GEOID == "39035" | GEOID == "39055" | GEOID == "39085" | 
                                        GEOID == "39103")$medincome, 
                               subset(acs_simple,
                                      GEOID == "39035" | GEOID == "39055" | GEOID == "39085" | 
                                        GEOID == "39103")$pop)

COL$pop <- subset(acs_simple, GEOID == "08041")$pop 
COL$medincome <- subset(acs_simple, GEOID == "08041")$medincome 

DEN$pop <- sum(subset(acs_simple, 
                      GEOID == "08005" |
                        GEOID == "08001" | 
                        GEOID == "08013" |
                        GEOID == "08014" |
                        GEOID == "08059" |
                        GEOID == "08031")$pop)
DEN$medincome <- weighted.mean(subset(acs_simple, 
                                      GEOID == "08005" |
                                        GEOID == "08001" | 
                                        GEOID == "08013" |
                                        GEOID == "08014" |
                                        GEOID == "08059" |
                                        GEOID == "08031")$medincome,
                               subset(acs_simple, 
                                      GEOID == "08005" |
                                        GEOID == "08001" | 
                                        GEOID == "08013" |
                                        GEOID == "08014" |
                                        GEOID == "08059" |
                                        GEOID == "08031")$pop)

DFW$pop <- sum(subset(acs_simple, 
                      GEOID == "48085" |
                        GEOID == "48113" | 
                        GEOID == "48139" |
                        GEOID == "48251" |
                        GEOID == "48439")$pop)
DFW$medincome <- weighted.mean(subset(acs_simple, 
                                      GEOID == "48085" |
                                        GEOID == "48113" | 
                                        GEOID == "48139" |
                                        GEOID == "48251" |
                                        GEOID == "48439")$medincome,
                               subset(acs_simple, 
                                      GEOID == "48085" |
                                        GEOID == "48113" | 
                                        GEOID == "48139" |
                                        GEOID == "48251" |
                                        GEOID == "48439")$pop)

FLG$pop <- subset(acs_simple, GEOID == "04005")$pop 
FLG$medincome <- subset(acs_simple, GEOID == "04005")$medincome 

GNV$pop <- subset(acs_simple, GEOID == "12001")$pop 
GNV$medincome <- subset(acs_simple, GEOID == "12001")$medincome 

LAW$pop <- subset(acs_simple, GEOID == "20045")$pop 
LAW$medincome <- subset(acs_simple, GEOID == "20045")$medincome 

LAX$pop <- sum(subset(acs_simple, 
                      GEOID == "06037" |
                        GEOID == "06065" | 
                        GEOID == "06071" |
                        GEOID == "06073")$pop)
LAX$medincome <- weighted.mean(subset(acs_simple, 
                                      GEOID == "06037" |
                                        GEOID == "06065" | 
                                        GEOID == "06071" |
                                        GEOID == "06073")$medincome,
                               subset(acs_simple, 
                                      GEOID == "06037" |
                                        GEOID == "06065" | 
                                        GEOID == "06071" |
                                        GEOID == "06073")$pop)
LCR$pop <- sum(subset(acs_simple, 
                      GEOID == "55063" |
                        GEOID == "27055" | 
                        GEOID == "27169")$pop)
LCR$medincome <- weighted.mean(subset(acs_simple, 
                                      GEOID == "55063" |
                                        GEOID == "27055" | 
                                        GEOID == "27169")$medincome,
                               subset(acs_simple, 
                                      GEOID == "55063" |
                                        GEOID == "27055" | 
                                        GEOID == "27169")$pop)

MOR$pop <- sum(subset(acs_simple, 
                      GEOID == "34013" |
                        GEOID == "34027" | 
                        GEOID == "34035" |
                        GEOID == "34039")$pop)
MOR$medincome <- weighted.mean(subset(acs_simple, 
                                      GEOID == "34013" |
                                        GEOID == "34027" | 
                                        GEOID == "34035" |
                                        GEOID == "34039")$medincome,
                               subset(acs_simple, 
                                      GEOID == "34013" |
                                        GEOID == "34027" | 
                                        GEOID == "34035" |
                                        GEOID == "34039")$pop)

MSP$pop <- sum(subset(acs_simple, 
                      GEOID == "27003" |
                        GEOID == "27053" | 
                        GEOID == "27123")$pop)
MSP$medincome <- weighted.mean(subset(acs_simple, 
                                      GEOID == "27003" |
                                        GEOID == "27053" | 
                                        GEOID == "27123")$medincome,
                               subset(acs_simple, 
                                      GEOID == "27003" |
                                        GEOID == "27053" | 
                                        GEOID == "27123")$pop)

NAS$pop <- sum(subset(acs_simple, 
                      GEOID == "47021" |
                        GEOID == "47037" | 
                        GEOID == "47149" |
                        GEOID == "47187")$pop)
NAS$medincome <- weighted.mean(subset(acs_simple, 
                                      GEOID == "47021" |
                                        GEOID == "47037" | 
                                        GEOID == "47149" |
                                        GEOID == "47187")$medincome,
                               subset(acs_simple, 
                                      GEOID == "47021" |
                                        GEOID == "47037" | 
                                        GEOID == "47149" |
                                        GEOID == "47187")$pop)

PHX$pop <- sum(subset(acs_simple, 
                      GEOID == "04013" |GEOID == "04021")$pop)
PHX$medincome <- weighted.mean(subset(acs_simple, 
                                      GEOID == "04013" |
                                        GEOID == "04021")$medincome,
                               subset(acs_simple, 
                                      GEOID == "04013" |
                                        GEOID == "04021")$pop) 

PTD$pop <- sum(subset(acs_simple, 
                      GEOID == "41005" |
                        GEOID == "53011" | 
                        GEOID == "27123")$pop)
PTD$medincome <- weighted.mean(subset(acs_simple, 
                                      GEOID == "41005" |
                                        GEOID == "53011" | 
                                        GEOID == "27123")$medincome,
                               subset(acs_simple, 
                                      GEOID == "41005" |
                                        GEOID == "53011" | 
                                        GEOID == "27123")$pop)

SC$pop <- subset(acs_simple, GEOID == "42027")$pop
SC$medincome <- subset(acs_simple, GEOID == "42027")$medincome

SD$pop <- subset(acs_simple, GEOID == "06073")$pop
SD$medincome <- subset(acs_simple, GEOID == "06073")$medincome

SF$pop <- sum(subset(acs_simple, 
                     GEOID == "06001" |
                       GEOID == "06013" | 
                       GEOID == "06041" |
                       GEOID == "06075"| 
                       GEOID == "06081"| 
                       GEOID == "06085")$pop)
SF$medincome <- weighted.mean(subset(acs_simple, 
                                     GEOID == "06001" |
                                       GEOID == "06013" | 
                                       GEOID == "06041" |
                                       GEOID == "06075"| 
                                       GEOID == "06081"| 
                                       GEOID == "06085")$medincome,
                              subset(acs_simple, 
                                     GEOID == "06001" |
                                       GEOID == "06013" | 
                                       GEOID == "06041" |
                                       GEOID == "06075" | 
                                       GEOID == "06081" | 
                                       GEOID == "06085")$pop)

SLC$pop <- sum(subset(acs_simple, 
                      GEOID == "49011" |
                        GEOID == "49035")$pop)
SLC$medincome <- weighted.mean(subset(acs_simple, 
                                      GEOID == "49011" |
                                        GEOID == "49035")$medincome,
                               subset(acs_simple, 
                                      GEOID == "49011" |
                                        GEOID == "49035")$pop)

SM$pop <- subset(acs_simple, GEOID == "48209")$pop 
SM$medincome <- subset(acs_simple, GEOID == "48209")$medincome 

SP$pop <- subset(acs_simple, GEOID == "12103")$pop 
SP$medincome <- subset(acs_simple, GEOID == "12103")$medincome 

WOO$pop <- subset(acs_simple, GEOID == "39169")$pop 
WOO$medincome <- subset(acs_simple, GEOID == "39169")$medincome 

#"Mahoning", "Trumbull"
YNG$pop <- sum(subset(acs_simple,
                      GEOID == "39099" | GEOID == "39155")$pop)
YNG$medincome <- weighted.mean(subset(acs_simple,
                                      GEOID == "39099" | GEOID == "39155")$medincome, 
                               subset(acs_simple,
                                      GEOID == "39099" | GEOID == "39155")$pop)

clusterLocations <- rbind(AA, ABQ, ATH, ATL, BEL, CED, CLE, COL, DEN, DFW, FLG, 
                          GNV, LAW, LAX, LCR, MOR, MSP, NAS, PHX, PTD, SC, 
                          SD, SF, SLC, SM, SP, WOO, YNG)

# Popdensity
clusterLocations$popdensity = clusterLocations$pop / 
  expanse(clusterLocations, unit = "km")

clusterLocations = as.data.frame(clusterLocations)
clusterLocations = subset(clusterLocations, 
                          select = c("id", "pop", "medincome", "popdensity"))

#we want to create a population density instead of just population, 
#so can do pop/(total_land/1000) for sqkm- bodies of water in regions generally aren't counted towards land total
multivariate <- merge(multivariate, clusterLocations, 
                          by.x = "cluster_location", by.y = "id")

# Water Use ---------------------------------------------------------------

# data from https://www.sciencebase.gov/catalog/item/get/5af3311be4b0da30c1b245d8, lightly cleaned
water <- read_excel("data/water.xlsx")
df <- water %>% 
  rename(GEOID = FIPS, 
         groundwater_total = 'TO-WGWFr', 
         surfacewater_total = 'TO-WSWFr', 
         fresh_total = 'TO-WFrTo') %>% 
  select(GEOID, 
         groundwater_total,
         surfacewater_total, 
         fresh_total
  )

## Freshwater total --------------------------------------------------------

multivariate$water_use <- NA
multivariate$water_use[multivariate$cluster_location == "Ann Arbor"] <- subset(df, GEOID == "26161")$fresh_total
multivariate$water_use[multivariate$cluster_location == "Albuquerque"] <- subset(df, GEOID == "35001")$fresh_total
multivariate$water_use[multivariate$cluster_location == "Athens"] <- sum(subset(df, 
                                                                                GEOID == "13219" |
                                                                                  GEOID == "13059")$fresh_total)
multivariate$water_use[multivariate$cluster_location == "Atlanta"] <- sum(subset(df, 
                                                                                 GEOID == "13121" |
                                                                                   GEOID == "13089" | 
                                                                                   GEOID == "13067")$fresh_total)
multivariate$water_use[multivariate$cluster_location == "Bellingham"] <- subset(df, GEOID == "53073")$fresh_total
multivariate$water_use[multivariate$cluster_location == "Cedar City"] <- subset(df, GEOID == "49021")$fresh_total
multivariate$water_use[multivariate$cluster_location == "Cleveland-Akron"] <- sum(subset(df, GEOID == "39035" |
                                                                                       GEOID == "39055" |
                                                                                       GEOID == "39085" | 
                                                                                       GEOID == "39103")$fresh_total)
multivariate$water_use[multivariate$cluster_location == "Colorado Springs"] <- subset(df, GEOID == "08041")$fresh_total
multivariate$water_use[multivariate$cluster_location == "Denver"] <- sum(subset(df, 
                                                                                GEOID == "08005" |
                                                                                  GEOID == "08001" | 
                                                                                  GEOID == "08013" |
                                                                                  GEOID == "08014" |
                                                                                  GEOID == "08059" |
                                                                                  GEOID == "08031")$fresh_total)
multivariate$water_use[multivariate$cluster_location == "Dallas Fort Worth"] <- sum(subset(df, 
                                                                                           GEOID == "48085" |
                                                                                             GEOID == "48113" | 
                                                                                             GEOID == "48139" |
                                                                                             GEOID == "48251" |
                                                                                             GEOID == "48439")$fresh_total)
multivariate$water_use[multivariate$cluster_location == "Flagstaff"] <- subset(df, GEOID == "04005")$fresh_total
multivariate$water_use[multivariate$cluster_location == "Gainesville"] <- subset(df, GEOID == "12001")$fresh_total
multivariate$water_use[multivariate$cluster_location == "Lawrence"] <- subset(df, GEOID == "20045")$fresh_total
multivariate$water_use[multivariate$cluster_location == "Los Angeles"] <- sum(subset(df, 
                                                                                     GEOID == "06037" |
                                                                                       GEOID == "06065" | 
                                                                                       GEOID == "06071" |
                                                                                       GEOID == "06073")$fresh_total)
multivariate$water_use[multivariate$cluster_location == "La Crosse"] <- sum(subset(df, 
                                                                                   GEOID == "55063" |
                                                                                     GEOID == "27055" | 
                                                                                     GEOID == "27169")$fresh_total)
multivariate$water_use[multivariate$cluster_location == "Morristown"] <- sum(subset(df, 
                                                                                    GEOID == "34013" |
                                                                                      GEOID == "34027" | 
                                                                                      GEOID == "34035" |
                                                                                      GEOID == "34039")$fresh_total)
multivariate$water_use[multivariate$cluster_location == "Minneapolis"] <- sum(subset(df, 
                                                                                     GEOID == "27003" |
                                                                                       GEOID == "27053" | 
                                                                                       GEOID == "27123")$fresh_total)
multivariate$water_use[multivariate$cluster_location == "Nashville"] <- sum(subset(df, 
                                                                                   GEOID == "47021" |
                                                                                     GEOID == "47037" | 
                                                                                     GEOID == "47149" |
                                                                                     GEOID == "47187")$fresh_total)
multivariate$water_use[multivariate$cluster_location == "Phoenix"] <- sum(subset(df, 
                                                                                 GEOID == "04013" |GEOID == "04021")$fresh_total)
multivariate$water_use[multivariate$cluster_location == "Portland"] <- sum(subset(df, 
                                                                                  GEOID == "41005" |
                                                                                    GEOID == "53011" | 
                                                                                    GEOID == "27123")$fresh_total)
multivariate$water_use[multivariate$cluster_location == "State College"] <- subset(df, GEOID == "42027")$fresh_total
multivariate$water_use[multivariate$cluster_location == "San Diego"] <- subset(df, GEOID == "06073")$fresh_total
multivariate$water_use[multivariate$cluster_location == "San Francisco"] <- sum(subset(df, 
                                                                                       GEOID == "06001" |
                                                                                         GEOID == "06013" | 
                                                                                         GEOID == "06041" |
                                                                                         GEOID == "06075" | 
                                                                                         GEOID == "06081" | 
                                                                                         GEOID == "06085")$fresh_total)
multivariate$water_use[multivariate$cluster_location == "Salt Lake City"] <- sum(subset(df, 
                                                                                        GEOID == "49011" |
                                                                                          GEOID == "49035")$fresh_total)
multivariate$water_use[multivariate$cluster_location == "San Marcos"] <- subset(df, GEOID == "48209")$fresh_total
multivariate$water_use[multivariate$cluster_location == "St Petersburg"] <- subset(df, GEOID == "12103")$fresh_total
multivariate$water_use[multivariate$cluster_location == "Wooster"] <- subset(df, GEOID == "39169")$fresh_total
multivariate$water_use[multivariate$cluster_location == "Youngstown"] <- sum(subset(df, 
                                                                                    GEOID == "39099" | 
                                                                                    GEOID == "39155")$fresh_total)

## Surfacewater Total ------------------------------------------------------

multivariate$surfacewater_total <- NA
multivariate$surfacewater_total[multivariate$cluster_location == "Ann Arbor"] <- subset(df, GEOID == "26161")$surfacewater_total
multivariate$surfacewater_total[multivariate$cluster_location == "Albuquerque"] <- subset(df, GEOID == "35001")$surfacewater_total
multivariate$surfacewater_total[multivariate$cluster_location == "Athens"] <- sum(subset(df, 
                                                                                         GEOID == "13219" |
                                                                                           GEOID == "13059")$surfacewater_total)
multivariate$surfacewater_total[multivariate$cluster_location == "Atlanta"] <- sum(subset(df, 
                                                                                          GEOID == "13121" |
                                                                                            GEOID == "13089" | 
                                                                                            GEOID == "13067")$surfacewater_total)
multivariate$surfacewater_total[multivariate$cluster_location == "Bellingham"] <- subset(df, GEOID == "53073")$surfacewater_total
multivariate$surfacewater_total[multivariate$cluster_location == "Cedar City"] <- subset(df, GEOID == "49021")$surfacewater_total
multivariate$surfacewater_total[multivariate$cluster_location == "Cleveland-Akron"] <- sum(subset(df, GEOID == "39035" |
                                                                                           GEOID == "39055" |
                                                                                           GEOID == "39085" | 
                                                                                           GEOID == "39103")$surfacewater_total)
multivariate$surfacewater_total[multivariate$cluster_location == "Colorado Springs"] <- subset(df, GEOID == "08041")$surfacewater_total
multivariate$surfacewater_total[multivariate$cluster_location == "Denver"] <- sum(subset(df, 
                                                                                         GEOID == "08005" |
                                                                                           GEOID == "08001" | 
                                                                                           GEOID == "08013" |
                                                                                           GEOID == "08014" |
                                                                                           GEOID == "08059" |
                                                                                           GEOID == "08031")$surfacewater_total)
multivariate$surfacewater_total[multivariate$cluster_location == "Dallas Fort Worth"] <- sum(subset(df, 
                                                                                                    GEOID == "48085" |
                                                                                                      GEOID == "48113" | 
                                                                                                      GEOID == "48139" |
                                                                                                      GEOID == "48251" |
                                                                                                      GEOID == "48439")$surfacewater_total)
multivariate$surfacewater_total[multivariate$cluster_location == "Flagstaff"] <- subset(df, GEOID == "04005")$surfacewater_total
multivariate$surfacewater_total[multivariate$cluster_location == "Gainesville"] <- subset(df, GEOID == "12001")$surfacewater_total
multivariate$surfacewater_total[multivariate$cluster_location == "Lawrence"] <- subset(df, GEOID == "20045")$surfacewater_total
multivariate$surfacewater_total[multivariate$cluster_location == "Los Angeles"] <- sum(subset(df, 
                                                                                              GEOID == "06037" |
                                                                                                GEOID == "06065" | 
                                                                                                GEOID == "06071" |
                                                                                                GEOID == "06073")$surfacewater_total)
multivariate$surfacewater_total[multivariate$cluster_location == "La Crosse"] <- sum(subset(df, 
                                                                                            GEOID == "55063" |
                                                                                              GEOID == "27055" | 
                                                                                              GEOID == "27169")$surfacewater_total)
multivariate$surfacewater_total[multivariate$cluster_location == "Morristown"] <- sum(subset(df, 
                                                                                             GEOID == "34013" |
                                                                                               GEOID == "34027" | 
                                                                                               GEOID == "34035" |
                                                                                               GEOID == "34039")$surfacewater_total)
multivariate$surfacewater_total[multivariate$cluster_location == "Minneapolis"] <- sum(subset(df, 
                                                                                              GEOID == "27003" |
                                                                                                GEOID == "27053" | 
                                                                                                GEOID == "27123")$surfacewater_total)
multivariate$surfacewater_total[multivariate$cluster_location == "Nashville"] <- sum(subset(df, 
                                                                                            GEOID == "47021" |
                                                                                              GEOID == "47037" | 
                                                                                              GEOID == "47149" |
                                                                                              GEOID == "47187")$surfacewater_total)
multivariate$surfacewater_total[multivariate$cluster_location == "Phoenix"] <- sum(subset(df, 
                                                                                          GEOID == "04013" |GEOID == "04021")$surfacewater_total)
multivariate$surfacewater_total[multivariate$cluster_location == "Portland"] <- sum(subset(df, 
                                                                                           GEOID == "41005" |
                                                                                             GEOID == "53011" | 
                                                                                             GEOID == "27123")$surfacewater_total)
multivariate$surfacewater_total[multivariate$cluster_location == "State College"] <- subset(df, GEOID == "42027")$surfacewater_total
multivariate$surfacewater_total[multivariate$cluster_location == "San Diego"] <- subset(df, GEOID == "06073")$surfacewater_total
multivariate$surfacewater_total[multivariate$cluster_location == "San Francisco"] <- sum(subset(df, 
                                                                                                GEOID == "06001" |
                                                                                                  GEOID == "06013" | 
                                                                                                  GEOID == "06041" |
                                                                                                  GEOID == "06075" | 
                                                                                                  GEOID == "06081" | 
                                                                                                  GEOID == "06085")$surfacewater_total)
multivariate$surfacewater_total[multivariate$cluster_location == "Salt Lake City"] <- sum(subset(df, 
                                                                                                 GEOID == "49011" |
                                                                                                   GEOID == "49035")$surfacewater_total)
multivariate$surfacewater_total[multivariate$cluster_location == "San Marcos"] <- subset(df, GEOID == "48209")$surfacewater_total
multivariate$surfacewater_total[multivariate$cluster_location == "St Petersburg"] <- subset(df, GEOID == "12103")$surfacewater_total
multivariate$surfacewater_total[multivariate$cluster_location == "Wooster"] <- subset(df, GEOID == "39169")$surfacewater_total
multivariate$surfacewater_total[multivariate$cluster_location == "Youngstown"] <- sum(subset(df, 
                                                                                    GEOID == "39099" | 
                                                                                      GEOID == "39155")$surfacewater_total)

## Groundwater total -------------------------------------------------------

multivariate$groundwater_total <- NA
multivariate$groundwater_total[multivariate$cluster_location == "Ann Arbor"] <- subset(df, GEOID == "26161")$groundwater_total
multivariate$groundwater_total[multivariate$cluster_location == "Albuquerque"] <- subset(df, GEOID == "35001")$groundwater_total
multivariate$groundwater_total[multivariate$cluster_location == "Athens"] <- sum(subset(df, 
                                                                                        GEOID == "13219" |
                                                                                          GEOID == "13059")$groundwater_total)
multivariate$groundwater_total[multivariate$cluster_location == "Atlanta"] <- sum(subset(df, 
                                                                                         GEOID == "13121" |
                                                                                           GEOID == "13089" | 
                                                                                           GEOID == "13067")$groundwater_total)
multivariate$groundwater_total[multivariate$cluster_location == "Bellingham"] <- subset(df, GEOID == "53073")$groundwater_total
multivariate$groundwater_total[multivariate$cluster_location == "Cedar City"] <- subset(df, GEOID == "49021")$groundwater_total
multivariate$groundwater_total[multivariate$cluster_location == "Cleveland-Akron"] <- sum(subset(df, GEOID == "39035" |
                                                                                                    GEOID == "39055" |
                                                                                                    GEOID == "39085" | 
                                                                                                    GEOID == "39103")$groundwater_total)
multivariate$groundwater_total[multivariate$cluster_location == "Colorado Springs"] <- subset(df, GEOID == "08041")$groundwater_total
multivariate$groundwater_total[multivariate$cluster_location == "Denver"] <- sum(subset(df, 
                                                                                        GEOID == "08005" |
                                                                                          GEOID == "08001" | 
                                                                                          GEOID == "08013" |
                                                                                          GEOID == "08014" |
                                                                                          GEOID == "08059" |
                                                                                          GEOID == "08031")$groundwater_total)
multivariate$groundwater_total[multivariate$cluster_location == "Dallas Fort Worth"] <- sum(subset(df, 
                                                                                                   GEOID == "48085" |
                                                                                                     GEOID == "48113" | 
                                                                                                     GEOID == "48139" |
                                                                                                     GEOID == "48251" |
                                                                                                     GEOID == "48439")$groundwater_total)
multivariate$groundwater_total[multivariate$cluster_location == "Flagstaff"] <- subset(df, GEOID == "04005")$groundwater_total
multivariate$groundwater_total[multivariate$cluster_location == "Gainesville"] <- subset(df, GEOID == "12001")$groundwater_total
multivariate$groundwater_total[multivariate$cluster_location == "Lawrence"] <- subset(df, GEOID == "20045")$groundwater_total
multivariate$groundwater_total[multivariate$cluster_location == "Los Angeles"] <- sum(subset(df, 
                                                                                             GEOID == "06037" |
                                                                                               GEOID == "06065" | 
                                                                                               GEOID == "06071" |
                                                                                               GEOID == "06073")$groundwater_total)
multivariate$groundwater_total[multivariate$cluster_location == "La Crosse"] <- sum(subset(df, 
                                                                                           GEOID == "55063" |
                                                                                             GEOID == "27055" | 
                                                                                             GEOID == "27169")$groundwater_total)
multivariate$groundwater_total[multivariate$cluster_location == "Morristown"] <- sum(subset(df, 
                                                                                            GEOID == "34013" |
                                                                                              GEOID == "34027" | 
                                                                                              GEOID == "34035" |
                                                                                              GEOID == "34039")$groundwater_total)
multivariate$groundwater_total[multivariate$cluster_location == "Minneapolis"] <- sum(subset(df, 
                                                                                             GEOID == "27003" |
                                                                                               GEOID == "27053" | 
                                                                                               GEOID == "27123")$groundwater_total)
multivariate$groundwater_total[multivariate$cluster_location == "Nashville"] <- sum(subset(df, 
                                                                                           GEOID == "47021" |
                                                                                             GEOID == "47037" | 
                                                                                             GEOID == "47149" |
                                                                                             GEOID == "47187")$groundwater_total)
multivariate$groundwater_total[multivariate$cluster_location == "Phoenix"] <- sum(subset(df, 
                                                                                         GEOID == "04013" |GEOID == "04021")$groundwater_total)
multivariate$groundwater_total[multivariate$cluster_location == "Portland"] <- sum(subset(df, 
                                                                                          GEOID == "41005" |
                                                                                            GEOID == "53011" | 
                                                                                            GEOID == "27123")$groundwater_total)
multivariate$groundwater_total[multivariate$cluster_location == "State College"] <- subset(df, GEOID == "42027")$groundwater_total
multivariate$groundwater_total[multivariate$cluster_location == "San Diego"] <- subset(df, GEOID == "06073")$groundwater_total
multivariate$groundwater_total[multivariate$cluster_location == "San Francisco"] <- sum(subset(df, 
                                                                                               GEOID == "06001" |
                                                                                                 GEOID == "06013" | 
                                                                                                 GEOID == "06041" |
                                                                                                 GEOID == "06075" | 
                                                                                                 GEOID == "06081" | 
                                                                                                 GEOID == "06085")$groundwater_total)
multivariate$groundwater_total[multivariate$cluster_location == "Salt Lake City"] <- sum(subset(df, 
                                                                                                GEOID == "49011" |
                                                                                                  GEOID == "49035")$groundwater_total)
multivariate$groundwater_total[multivariate$cluster_location == "San Marcos"] <- subset(df, GEOID == "48209")$groundwater_total
multivariate$groundwater_total[multivariate$cluster_location == "St Petersburg"] <- subset(df, GEOID == "12103")$groundwater_total
multivariate$groundwater_total[multivariate$cluster_location == "Wooster"] <- subset(df, GEOID == "39169")$groundwater_total
multivariate$groundwater_total[multivariate$cluster_location == "Youngstown"] <- sum(subset(df, 
                                                                                             GEOID == "39099" | 
                                                                                               GEOID == "39155")$groundwater_total)



# Write it -----------------------------------------------------

write.csv(multivariate, "data/multivariate.csv")
